refactor: refactor structure order
