+++ app/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { Header } from '../components/header'
-import { Banner } from '../components/content'
+import { Banner } from '../components/content/Banner'
 
 export default function Page() {
   return (
