+++ components/content/Banner/Banner.module.sass
@@ -40,11 +40,21 @@
     padding: 1rem
 
 .button
+    align-items: center
     background: var(--button-gradient, #00baff)
     border: none
     border-radius: .5rem
     color: var(--white, #ffffff)
     cursor: pointer
+    display: flex
     font-weight: bold
     font-size: 1rem
+    gap: .5rem
+    justify-content: center
     padding: 1rem
+
+    &--icon
+        margin-top: -.3rem
+        width: .8rem
+
+
