+++ components/content/Banner/index.ts
